requirements.txt
numpy
pandas
tensorflow
keras
matplotlib
Pillow
flask